make lodestone a lib
